Se agregaron las funcionalidades de ProductController

+++ GymTEC/Data_input_models/manage_product/Data_input_product.cs
@@ -6,6 +6,6 @@
         public string product_name { get; set; }      // nombre
         public string barcode { get; set; }           // código de barras
         public string description { get; set; }       // descripción
-        public double cost { get; set; }              // costo
+        public int cost { get; set; }              // costo
     }
 }
